requirements.txt
asgiref==3.7.2
Django==5.0.1
django-auth-ldap==4.6.0
ldap3==2.9.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-decouple==3.8
python-ldap==3.4.4
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
